backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-generativeai==0.3.1
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
langdetect==1.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
